// React/src/interfaces/ITrackerItem.ts
import { IBase } from './IBase';
import { IBaseWithName } from './IBaseWithName';
import { IBusinessUnit } from './IBusinessUnit';
import { ILocation } from './ILocation';
import { ITrackerQuestion } from './ITrackerQuestion';
import { TQuestionValue } from './TQuestionValue';

export interface ITrackerItem extends IBase {
  name: string;
  description?: string;
  categoryId: string;
  regulatoryBodyId: string;
  dueDate: Date;
  dueDateCalculation: 'fromDueDate' | 'fromCompletionDate';
  dueDateEditable?: boolean;
  frequency: string;
  businessUnitsIds: string[];
  evidenceItems: string[];
  allowAttachments?: boolean;
  questions?: ITrackerQuestion<TQuestionValue>[];
  locationsIds?: string[];
  published: boolean;
  reference?: string;

  locations?: ILocation[];
  category?: IBaseWithName;
  regulatoryBody?: IBaseWithName;
  businessUnits?: IBusinessUnit[];
}